+++ src/code42cli/profile/password.py
@@ -9,30 +9,32 @@ from code42cli.profile.config import get_config_accessor, ConfigAccessor
 _ROOT_SERVICE_NAME = u"code42cli"
 
 
-def get_password(profile_name):
-    """Gets your currently stored password for your profile."""
-    accessor = get_config_accessor()
-    profile = accessor.get_profile(profile_name)
-    service_name = _get_service_name(profile_name)
+def get_stored_password(profile_name):
+    """Gets your currently stored password for the given profile name."""
+    profile = _get_profile(profile_name)
+    service_name = _get_service_name(profile.name)
     username = _get_username(profile)
     password = keyring.get_password(service_name, username)
     return password
 
 
-def set_password_from_prompt(profile_name):
-    """Prompts and sets your password for your profile."""
-    password = getpass()
-    accessor = get_config_accessor()
-    profile = accessor.get_profile(profile_name)
-    service_name = _get_service_name(profile_name)
-    username = _get_username(profile)
-    keyring.set_password(service_name, username, password)
-    print(u"'Code42 Password' updated.")
-    return password
-
-
 def get_password_from_prompt():
+    """Prompts you and returns what you input."""
     return getpass()
+
+
+def set_password(profile_name, new_password):
+    """Sets your password for the given profile name."""
+    profile = _get_profile(profile_name)
+    service_name = _get_service_name(profile.name)
+    username = _get_username(profile)
+    keyring.set_password(service_name, username, new_password)
+    print(u"'Code42 Password' updated.")
+
+
+def _get_profile(profile_name):
+    accessor = get_config_accessor()
+    return accessor.get_profile(profile_name)
 
 
 def _get_service_name(profile_name):