+++ src/code42cli/util.py
@@ -27,13 +27,13 @@ def get_user_project_path(subdir=""):
 
 def open_file(file_path, mode, action):
     """Wrapper for opening files, useful for testing purposes."""
-    with open(file_path, mode) as f:
+    with open(file_path, mode, encoding="utf-8") as f:
         action(f)
 
 
 def print_error(error_text):
     """Prints red text."""
-    print("\033[91mUSAGE ERROR: {}\033[0m".format(error_text))
+    print("\033[91mERROR: {}\033[0m".format(error_text))
 
 
 def print_bold(bold_text):